Add a new structure: zone register
Zone register is used for storing pointers to zones and their
corresponding LDAP distinguished names. These can then be accessed by
dns name of the zone. The data is needed for conversion of a dns name to
DN and to get a pointer of a zone when we need to make changes to it. We
could use dns_view_findzone() for this, but that way we would not have
any assurance that the found zone is really managed by us.