Use the zone register for internal storage of zones
Moving the zone registration into a separate file will make the code
more modular and easier to read and change. Also, we are preparing the
refresh_zones_from_ldap() function for it's intended purpose of being
able to add zones and later modify them if something has changed.
Unfortunately, we will only be able to change some of the zone
attributes (for now the update policy). We will not be able to create
new zones as they are added to LDAP because that requires a reload of
the name server.