Gracefully fail if no zones are found in the initial lookup
If the initial search fails to find any zones, destroy the database
instance, but return ISC_R_SUCCESS to BIND, since this is not
technically a fault because of which BIND should abort. We do this
because there is no point in refreshing zones if there are not any and
it is not possible to add them later.