Add support for write-back to LDAP after a DNS update.
Current code doesn't try to do re-synchronization after an error
during update. The error is logged and SERVFAIL is returned to
the client but no re-synchronization is scheduled.

Some hardening against race conditions will be required. Some data could
be potentially lost if events in syncrepl queue are not processed fast
enough.

Signed-off-by: Petr Spacek <pspacek@redhat.com>