Use RBTDB instead of internal LDAP cache.
Internal RBTDB is created for each zone and most calls from
internal database API are relayed to the internal RBTDB.

Each RBTDB initially contains fake SOA and NS records and
the content of the internal RBTDB is synchronized with LDAP
during start-up.
Responses retured to clients during initial synchronization
can be incomplete and can contain illegal SOA and NS records.

DNS updates are not propagated to LDAP and SOA serial number is not
maintained at all.

LDAP MODRDN operation is not supported because current code lacks
LDAP entryUUID->DNS name mapping database.

Side-effect of transition to RBTDB is that DNS wildcards are supported
from now. Another side-effect is that non-existing records
are virtually 'cached' because RBTDB is authoritative source
of information. As a result, read-performance is nearly same as
with plain BIND 9.

https://fedorahosted.org/bind-dyndb-ldap/ticket/95
https://fedorahosted.org/bind-dyndb-ldap/ticket/6

Signed-off-by: Petr Spacek <pspacek@redhat.com>