Rework locking around DNS views.
It turned out that previous usage of isc_task_beginexclusive() was
completely incorrect and didn't actually lock anything. As a result the
plugin was randomly crashing.

Code was reworked in a way that isc_task_beginexclusive() is always called on
inst->task and all isc_task_beginexclusive() calls are made from inside
of inst->task.

See comments in lock.c for more details.

Signed-off-by: Petr Spacek <pspacek@redhat.com>