Implement handling of inactive master zones.
Inactive zones are loaded into memory in the same way as active zones.
The only difference is that inactive zones are not added to DNS view
used by bind-dyndb-ldap.

Zone will be re-added to DNS view if idnsActiveZone attribute is changed
to TRUE so the change should be almost immediate.

Usual zone maintenance (serial number maintenance, DNSSEC in-line signing etc.)
is done for all zones, no matter if the zone is active or not.
This allows us to maintain zone journal so IXFR works correctly even
after zone re-activation.

Other option is to blow all DNS databases and do full reload from LDAP
whenever idnsZoneActive changes in single zone. It doesn't seem like a
good idea.

More fine-grained approach where only single zone is reloaded is not possible
at the moment because we lack UUID database which is necessary for correct
SyncRepl reinitialization/stop-and-resume.

https://fedorahosted.org/bind-dyndb-ldap/ticket/127

Signed-off-by: Petr Spacek <pspacek@redhat.com>