SyncPTR: Read PTR record values from RBTDB instead of LDAP.
Reading values directly from LDAP was racy because the PTR record value can be
modified between time-of-check and time-of-use.

This change is not fixing the race completely, SyncPTR needs to be split
to separate event which will go throught the same event queue as other
updates to the reverse zone.

https://fedorahosted.org/bind-dyndb-ldap/ticket/155