Unload automatic empty zones only if conflicting forward zone has policy 'only'.
This is exactly what BIND does so now we are compatible with BIND
semantics.

A warning message is printed when an conflicting forward zone with
policy 'first' is configured. This mimics behavior implemented in
BIND RT #41441 for BIND 9.11.

Unfortunatelly this is incompatible change when compared to
bind-dyndb-ldap 3.3+ which was unloading zones without regard to the forwarding
policy. On the other hand, bind-dyndb-ldap 3.3+ was unloading only some
subset of empty zones so its behavior was broken anyway.

During upgrade users should to convert all forward zones which are
conflicting with automatic empty zones to forward policy 'only'.

Master zones are not affected by this change.

https://fedorahosted.org/bind-dyndb-ldap/ticket/160