Untangle forwarder configuration from LDAP code.
Previously code handling forwarders was so interleaved with LDAP code
that it was impossible to add another layer for per-server config.

This is preparatory work which untangles LDAP and forwarders to enable
further development.

As far as I can tell the user-visible behavior should not be changed by
this patch.

https://fedorahosted.org/bind-dyndb-ldap/ticket/162