Remove misleading zonemgr check from task selection mechanism.
It seems that there is no way how a zone could end up in zone register
without having zone->task.

Theoreticallly the check should be no-op but I want to be sure that
fallback is not used for existing zones.

This is an attempt to make task selection deterministic so
re-synchronization can be free of race conditions caused by task
selection mechanism.

https://fedorahosted.org/bind-dyndb-ldap/ticket/125