spec: remove unnecessary bind-pkcs11 dependency
The bind-pkcs11 dependency was originally introduced in Fedora
specfile in 2014 to help with ipa-server-upgrade. This was a
temporary workaround. Since Fedora life cycle is 13 months and
this change happened more than two years ago, it is safe to
remove the workaround.

bind-dyndb-ldap does not actually depend on bind-pkcs11, although it
is a dependency that is enforced by freeipa.